% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite_matching_collapsed.R
\name{bipartite_matching_collapsed}
\alias{bipartite_matching_collapsed}
\title{Bipartite to unipartite matching via collapsing the bipartite network}
\usage{
bipartite_matching_collapsed(
  A,
  B,
  collapsed_method = c("omp", "corr", "cov", "glasso", "mb"),
  seeds = NULL
)
}
\arguments{
\item{A}{Adjacency matrix of the unipartite graph.}

\item{B}{Bipartite graph incidence matrix, where rows correspond to the vertices in common
with the unipartite graph.}

\item{collapsed_method}{Collapsed representation for the bipartite graph. The options are one-mode projection (\texttt{omp}),
correlation (\texttt{corr}), covariance (\texttt{cov}), graphical lasso (\texttt{glasso}) or nodewise regression (\texttt{mb}).}

\item{seeds}{If some vertices have known correspondence, a vector containing the indexes of these
vertices can be passed through this parameter, and the corresponding rows of A and B are assumed to be aligned.
The algorithm will then match the remaining vertices. The default is NULL if no seeds are available.}
}
\value{
A permutation Q that approximately minimizes the difference between the edges of A and the collapsed B.
}
\description{
This function performs graph matching between a bipartite and a unipartite graphs that are assumed to
share a common set of vertices. The method performs an alignment of the edges of the unipartite graphs
with the edges of a collapsed bipartite graph into a unipartite, following the collapsed representation
selected.
}
