% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrafunctions.R
\name{edge_error}
\alias{edge_error}
\title{Calculate edge error}
\usage{
edge_error(Q1, Q2, A)
}
\arguments{
\item{Q1}{Permutation matrix}

\item{Q2}{Permutation matrix}

\item{A}{adjacency matrix}
}
\value{
Proportion of edges with different matching.
}
\description{
This function counts the number edge disagreements between A and a permuted version of it, and
reports the proportion of edges that are different.
If the two oermutations are the same, then the result is zero.
}
