% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrafunctions.R
\name{gm_error}
\alias{gm_error}
\title{Calculate graph matching error}
\usage{
gm_error(Q1, Q2)
}
\arguments{
\item{Q1}{Permutation matrix}

\item{Q2}{Permutation matrix}
}
\value{
Proportion of vertices with different matching.
}
\description{
This function counts the number of vertices that are incorrectly matched across two permutations, and divides
over the total number of vertices.
}
