% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrafunctions.R
\name{graphical_model_errors}
\alias{graphical_model_errors}
\title{Calculate errors in estimating edges of a graphical model}
\usage{
graphical_model_errors(Theta, A, Q = diag(ncol(A)))
}
\arguments{
\item{Theta}{A matrix where non-zeros indicate edges.}

\item{A}{Adjacency matrix of the graphical model}

\item{Q}{Alignment permutation matrix between A and Theta. The default is the identity matrix.}
}
\value{
False positive rate (FPR), false negative rate (FNR) and total support error.
}
\description{
This function reports the false positive rate, false negative rate, and proportion of errors for
estimating the edges of a graphical model.
}
